requirements.txt
streamlit

google-generativeai

python-dotenv

langchain

PyPDF2

chromadb

faiss-cpu

langchain_google_genai

regex

scikit-learn

langchain_core

spacy

python-docx

transformers

sentence-transformers

spacy

docx2txt

torch

ctransformers

langchain

langchain-community

langchain-core

docx
 